Update README.md
The Driver Drowsiness Detection System uses Python, OpenCV, and Deep Learning to detect driver fatigue in real-time. It analyzes facial landmarks on the eyes, mouth, and face to identify signs of drowsiness, such as eye closure and yawning, and triggers alerts. Sample images included in this repository show how facial landmarks are detected and processed. The system leverages pre-trained models but can be improved with custom datasets for better accuracy, making it a scalable solution for road safety.